requirements.txt
python=3.10
arxiv
pdfplumber
PyMuPDF
pillow
requests
